pyproject.toml
[project]
name = "agent-performance-evaluator"
version = "0.1.0"
description = "A comprehensive benchmarking framework for evaluating AI agent implementations and their tool-calling capabilities across multiple scenarios."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.24",
    "langchain-anthropic>=0.3.12",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "llama-index>=0.12.34",
    "llama-index-llms-anthropic>=0.6.10",
    "openai-agents>=0.0.14",
    "pandas>=2.2.3",
    "py-calling-agent[litellm]>=0.3.1",
    "tqdm>=4.67.1",
]
